Vendor some elisp packages directly into the repo
I will add some things, and their dependencies, here if they cannot be
found on any of the usual package repos.

The reason for these packages is that I want crosshairs.el, but
hl-line+ is not on any of the usual package repos. Quite strange,
since crosshairs which depends on hl-line+ is on marmalade (however it
is an old version). To solve this mess, just include crosshairs.el +
dependencies directly into my emacs conf repo.

In the future: If I can find them in a git repo, but not on
melpa, maybe I'll use a git submodule.