Vendor prolog.el
This is a specific version of the prolog mode that I startd to use
some time during my studies at KTH. I remember preferring this one to
the one included in emacs for some reason. It might simply be that it
supported sicstus prolog better.

TODO: compare running with/without this for different versions of prolog