use-package-extract: clean up ensure condition
Ignore the misfeature of use-package that :ensure can work without any
"argument". Handling that properly is too messy. For me, this just
means that I have to be sure to always hand an argument to :ensure in
my use-package declarations in init.el, which shouldn't be too bad.

TODO: invent better syntax for use-package