Add bitbake-el as a submodule
This package doesn't seem to exist in the stable melpa repo (it does
in the unstable one, however) so let's include it like a submodule for
now.

EDIT: Also I had to fix some regexes in it etc., so now I ended up
with my own fork of this repo.