package-settings: Switch to using unstable melpa
stable melpa is for wusses

real reason: shader-mode doesn't seem to exist in stable melpa. might
be the case for more packages.

TODO: Maybe get rid of the gnu and ELPA archives in the future. I'm
not sure if I use any packages that are only available there any more.