init.el: Remove :ensure t from csharp-mode use-package definition
Because in emacs 29 csharp-mode is included in the default emacs
distribution, so let's avoid accidentally installing it from melpa or
so like perhaps some of my scripts might do accidentally.

Unfortunately means I am forced to manually install it on some systems
where I am still stuck on emacs 28 or 27.