Refactoring of my various ways of adding things to auto-mode-alist
Now I use :mode from use-package everywhere, which is useful I guess.
I still think using use-package for built-in packages (and to some
extent I guess my vendored ones?) feels a bit off though.

Note that the makefile one is removed completely as that setting as I
think that has been a default for a long while back.

Also I took the opportunity to refactor the eval-after-load for
prolog-mode to also use use-package.