init.el: Predicate the linum-mode alias on using emacs 29 or newer
linum-mode was only removed with emacs 29. in the older ones that have
both display-line-numbers-mode and linum-mode let's just leave it as
is, even though it might be nicer to use the newer one in some cases
there too.

That could be done in the future with some more complex
logic to handle the three cases of:
1. linum-mode only,
2. linum-mode and display-line-numbers-mode
3. display-line-numbers-mode only
But I doubt it's really neccessary for my usecase yet