Remove the rebinding of RET to newline-and-indent for newer than emacs 24
Since the we have electric newline mode (which is on by default) which
seems to mostly do what I want.

There are some problems though when using lsp-mode with at least C++
projects where C-j seems to insert a newline, putting cursor at start
of it, but it also inserts whitespace after the cursor... So another
possible route might be to ditch electric newline mode and just keep
rebinding like this. Who knows...