I created some silly functions to loop over all elisp functions
I made this mainly because I wanted a quick hack to get an overview of
all native eshell functions, as you can get with
list-eshell-functions.

I guess in the future I could make an eshell built in function that
lists them all inside eshell or something. Maybe even with
descriptions or so. eshell is kinda lacking a built-in help function