I guess let's finally switch iswitchb-mode out for icomplete-mode
icomplete-mode + fido-mode (or fido-vertical-mode) seems pretty
decent. I could probably replace my few uses of helm with this too (TODO)

I'm not sure if I'll like it for find-file etc. though. Try and hack
together something that makes icomplete-mode not work for find-file.
Or perhaps make it work in regular icomplete-mode for
that? (potentially confusing)