No need any longer to keep a submodule for vterm. Instead just use it from ELPA
We switch to just using it from elpa (or if installed with the distros
package manager).

I think back in the day it wasn't available from either, but I just
kept on using the submodule even though it was no longer neccessary.

TODO: Reconsider how i-have-vterm is defined as well. I think at this
point it should more-or-less be enough to be on not-windows.