Fix use-package declaration for powershell
My attempt at being explicit about loading it eagerly seems to have
failed, which in turn causes the use-package declaration for eglot to
fail. Which actually causes my customizations for eglot to not get
implied, such as the very-important-for-my-sanity setting of
eglot-ignored-server-capabilities to
'(:documentOnTypeFormattingProvider)

Apparently the :after in (use-package eglot ...) causes none of that
to run even when eglot gets forcibly loaded by running any of the
eglot commands or just running plain (require 'eglot). Seems like the
:after declaration is a bit too powerful... I just want it to prevent
it to autoload without the other...

On that note maybe there should be a sort of :requires modeled on the
systemd equivalent so that loading eglot will force-load powershell if
it hasn't been loaded already. Although this naming does end up a bit
stupid and implies that eglot cannot function without powershell,
which isn't really true...

TODO: Check whether the need to load eglot after the powershell mode
is still a neccessity?